Initial implementation of P2P tx distribution
The first iteration of P2P networking here is based on pea2pea and its
Noise example.

The transaction handling is fully built into the P2P code, but should be
easily refactored to proper PubSub model later on.

This is very much MVP right now, but should provide some ideas to build
on later.