Restructure `Cargo.toml` for the node crate (#101)
Node crate contains both, the node binary, but also the library for
writing e.g. JSON-RPC clients or building transactions out of the node.

In order to avoid certain dependencies in the library (most notably
the vsock related ones), those problematic dependencies are marked as
optional and moved behind a feature flag that the binary build depends
on.