Add flag for `--no-execution` to disable tx exec
This change allows running a full node that doesn't execute transactions.

Transaction execution is disabled in two ways:
- Mempool won't populate its `deque` with transactions, but only
  persists them to the storage. This way, transactions will never
  become available for Scheduler.

- Provider is configured with `NoExecProvider` implementation, that is
  complete no-op implementation for VM execution.

With these changes, a node can be run in such a way that it provides
full functionality to the network but doesn't require special HW. Main
use case for this is a combined JSON-RPC + Archive node instance, that
can be also used as a P2P bootstrap node.