Add get tx execution output rpc (#112)
* add rpc getTxExecutionOutput

* test + add url scheme

* change cli cargo.toml to build on Mac

* Mark transaction as executed (#114)

* Log P2P connect error (#115)

Ignoring the `node.connect()` error hides a problem if e.g. the peer
doesn't allow connection in to advertised P2P address.

* Simplify build: Invert the node-binary feature flag (#117)

This change allows building everything from the workspace by-default,
but allows reduced dependencies when only building `gevulot-cli`.

* Handle Noise codec error (#116)

In some cases, the handshake state might be in wrong sequence, which
generates an error. This change adds proper error handling so that the
thread doesn't panic.

* Improve txfile URL handling (#118)

The address generated for txfile URL originates from the P2P bind
address. In case the node is behind NAT, the address should take this
into account.

* Change get-tx return to a json tx

* add rpc getTxExecutionOutput

* test + add url scheme

* Change get-tx return to a json tx

* add rpc getTxExecutionOutput

* test + add url scheme

* Change get-tx return to a json tx

---------

Co-authored-by: Tuomas Mäkinen <1947505+tuommaki@users.noreply.github.com>