Prefix podman commands with `sudo` for now (#125)
Currently the Gevulot container is spawned from `root` when launched
by systemd. In Podman, container registries are per-user by default so
when pulling the latest version of container, the further container
operations must be done using the same user account as well. This is
especially important when there are critical changes in the container
such as database migrations. Therefore podman commands here should be
prefixed with `sudo` in order to match the correct container catalog.