Add support for P2P protocol v1 (when orig. is v0) (#175)
This change makes next step in P2P protocol versioning support. While
at it, this also takes the first step to add metrics into v1 protocol.

The prometheus metrics export is not part of this change yet, though.

The protocol versioning change here is built to be backwards compatible.
Message versions default to v0, but have support for optional v1
compatibility as well. This change does not yet have support for
protocol version negotiation between nodes, though it might be that this
won't be even necessary, if nodes upgrade fast enough.