fix: dynamic QEMU path resolution
Prior to this commit, it is assumed that the QEMU binary resides within
the canonical Linux path. However, certain systems like NixOS may place
QEMU in a local user folder instead of the global `/usr/bin` path,
causing an error when attempting to locate the QEMU binary.

This change incorporates a `which` command, which searches through the
current `PATH` environment variable and identifies the location of the
QEMU binary.